Changed project folder structure: opentelemetry, opentracing, opentelemetry-example, opentracing-example as project/folders
