+++ build.sbt
@@ -40,7 +40,7 @@ lazy val root =
   project
     .in(file("."))
     .settings(skip in publish := true)
-    .aggregate(openTracing, openTelemetry, example)
+    .aggregate(opentracing, opentelemetry, opentracingExample, opentelemetryExample)
 
 val http4sVersion        = "0.21.3"
 val jaegerVersion        = "1.2.0"
@@ -50,9 +50,9 @@ val opentelemetryVersion = "0.3.0"
 val zipkinVersion        = "2.12.3"
 val zioVersion           = "1.0.0-RC18-2"
 
-lazy val openTracing =
+lazy val opentracing =
   project
-    .in(file("modules/opentracing"))
+    .in(file("opentracing"))
     .settings(stdSettings("zio-opentracing"))
     .settings(
       libraryDependencies := Seq(
@@ -66,9 +66,9 @@ lazy val openTracing =
       )
     )
 
-lazy val openTelemetry =
+lazy val opentelemetry =
   project
-    .in(file("modules/opentelemetry"))
+    .in(file("opentelemetry"))
     .settings(stdSettings("zio-opentelemetry"))
     .settings(
       libraryDependencies := Seq(
@@ -83,31 +83,45 @@ lazy val openTelemetry =
 
 Global / testFrameworks := Seq(new TestFramework("zio.test.sbt.ZTestFramework"))
 
-lazy val example =
+lazy val exampleDependencies = Seq(
+  "org.typelevel"                %% "cats-core"                     % "2.1.1",
+  "io.circe"                     %% "circe-generic"                 % "0.13.0",
+  "org.http4s"                   %% "http4s-core"                   % http4sVersion,
+  "org.http4s"                   %% "http4s-blaze-server"           % http4sVersion,
+  "org.http4s"                   %% "http4s-dsl"                    % http4sVersion,
+  "org.http4s"                   %% "http4s-circe"                  % http4sVersion,
+  "io.jaegertracing"             % "jaeger-core"                    % jaegerVersion,
+  "io.jaegertracing"             % "jaeger-client"                  % jaegerVersion,
+  "io.jaegertracing"             % "jaeger-zipkin"                  % jaegerVersion,
+  "com.github.pureconfig"        %% "pureconfig"                    % "0.12.3",
+  "com.softwaremill.sttp.client" %% "async-http-client-backend-zio" % sttpVersion,
+  "com.softwaremill.sttp.client" %% "circe"                         % sttpVersion,
+  "dev.zio"                      %% "zio-interop-cats"              % "2.0.0.0-RC12"
+)
+
+lazy val opentracingExample =
   project
-    .in(file("modules/example"))
-    .settings(stdSettings("example"))
+    .in(file("opentracing-example"))
+    .settings(stdSettings("opentracing-example"))
     .settings(skip in publish := true)
     .settings(
-      libraryDependencies := Seq(
-        "org.typelevel"                %% "cats-core"                     % "2.1.1",
-        "io.circe"                     %% "circe-generic"                 % "0.13.0",
-        "org.http4s"                   %% "http4s-core"                   % http4sVersion,
-        "org.http4s"                   %% "http4s-blaze-server"           % http4sVersion,
-        "org.http4s"                   %% "http4s-dsl"                    % http4sVersion,
-        "org.http4s"                   %% "http4s-circe"                  % http4sVersion,
-        "io.jaegertracing"             % "jaeger-core"                    % jaegerVersion,
-        "io.jaegertracing"             % "jaeger-client"                  % jaegerVersion,
-        "io.jaegertracing"             % "jaeger-zipkin"                  % jaegerVersion,
-        "com.github.pureconfig"        %% "pureconfig"                    % "0.12.3",
-        "com.softwaremill.sttp.client" %% "async-http-client-backend-zio" % sttpVersion,
-        "com.softwaremill.sttp.client" %% "circe"                         % sttpVersion,
-        "dev.zio"                      %% "zio-interop-cats"              % "2.0.0.0-RC12",
-        "io.zipkin.reporter2"          % "zipkin-reporter"                % zipkinVersion,
-        "io.zipkin.reporter2"          % "zipkin-sender-okhttp3"          % zipkinVersion,
-        "io.opentelemetry"             % "opentelemetry-exporters-jaeger" % opentelemetryVersion,
-        "io.opentelemetry"             % "opentelemetry-sdk"              % opentelemetryVersion,
-        "io.grpc"                      % "grpc-netty-shaded"              % "1.28.0"
+      libraryDependencies := exampleDependencies ++ Seq(
+        "io.zipkin.reporter2" % "zipkin-reporter"       % zipkinVersion,
+        "io.zipkin.reporter2" % "zipkin-sender-okhttp3" % zipkinVersion
       )
     )
-    .dependsOn(openTracing, openTelemetry)
+    .dependsOn(opentracing)
+
+lazy val opentelemetryExample =
+  project
+    .in(file("opentelemetry-example"))
+    .settings(stdSettings("opentelemetry-example"))
+    .settings(skip in publish := true)
+    .settings(
+      libraryDependencies := exampleDependencies ++ Seq(
+        "io.opentelemetry" % "opentelemetry-exporters-jaeger" % opentelemetryVersion,
+        "io.opentelemetry" % "opentelemetry-sdk"              % opentelemetryVersion,
+        "io.grpc"          % "grpc-netty-shaded"              % "1.28.0"
+      )
+    )
+    .dependsOn(opentelemetry)
