--- modules/example/src/main/scala/zio/telemetry/example/config/package.scala
@@ -1,9 +1,0 @@
-package zio.telemetry.example
-
-import zio.Has
-
-package object config {
-
-  type Configuration = Has[Configuration.Service]
-
-}
