--- modules/example/src/main/scala/zio/telemetry/example/http/package.scala
@@ -1,10 +1,0 @@
-package zio.telemetry.example
-
-import zio.ZIO
-import zio.clock.Clock
-
-package object http {
-
-  type AppTask[A] = ZIO[Clock, Throwable, A]
-
-}
