--- modules/example/src/main/scala/zio/telemetry/opentelemetry/config/package.scala
@@ -1,9 +1,0 @@
-package zio.telemetry.opentelemetry
-
-import zio.Has
-
-package object config {
-
-  type Configuration = Has[Configuration.Service]
-
-}
