+++ opentelemetry-example/src/main/resources/application.conf
