+++ opentelemetry-example/src/main/scala/zio/telemetry/opentelemetry/tracing/example/config/package.scala
@@ -1,0 +1,9 @@
+package zio.telemetry.opentelemetry.tracing.example
+
+import zio.Has
+
+package object config {
+
+  type Configuration = Has[Configuration.Service]
+
+}
