+++ opentracing-example/src/main/resources/application.conf
