+++ opentracing-example/src/main/scala/zio/telemetry/opentracing/example/config/Config.scala
@@ -1,0 +1,17 @@
+package zio.telemetry.opentracing.example.config
+
+import zio.telemetry.opentracing.example.config.Config._
+
+final case class Config(proxy: ProxyConfig, backend: BackendConfig, tracer: TracerHost)
+
+object Config {
+
+  final case class ProxyConfig(host: String, port: Int)
+
+  final case class BackendUrl(url: String) extends AnyVal
+
+  final case class BackendConfig(host: String, port: Int)
+
+  final case class TracerHost(host: String) extends AnyVal
+
+}
