+++ opentracing-example/src/main/scala/zio/telemetry/opentracing/example/config/package.scala
@@ -1,0 +1,9 @@
+package zio.telemetry.opentracing.example
+
+import zio.Has
+
+package object config {
+
+  type Configuration = Has[Configuration.Service]
+
+}
