+++ opentracing-example/src/main/scala/zio/telemetry/opentracing/example/http/Statuses.scala
@@ -1,0 +1,11 @@
+package zio.telemetry.opentracing.example.http
+
+import io.circe._
+import io.circe.generic.semiauto._
+
+final case class Statuses(data: List[Status]) extends AnyVal
+
+object Statuses {
+  implicit val decoder: Decoder[Statuses] = deriveDecoder
+  implicit val encoder: Encoder[Statuses] = deriveEncoder
+}
