+++ opentracing-example/src/main/scala/zio/telemetry/opentracing/example/http/package.scala
@@ -1,0 +1,10 @@
+package zio.telemetry.opentracing.example
+
+import zio.ZIO
+import zio.clock.Clock
+
+package object http {
+
+  type AppTask[A] = ZIO[Clock, Throwable, A]
+
+}
